pyproject.toml
[project]
name = "things3-blade-mcp"
version = "0.1.2"
description = "Best-in-class MCP server for Things 3 — token-efficient, full CRUD, AppleScript + URL scheme"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [{name = "Piers"}]
keywords = ["things3", "mcp", "task-management", "gtd", "applescript"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Scheduling",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "fastmcp>=2.0.0",
    "things-py>=0.0.15",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
things3-blade-mcp = "things3_blade_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/things3_blade_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
